requirements.txt
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
fastapi==0.118.0
-e git+https://github.com/Orange-Hack/finam-x-hse-trade-ai-hack-trader.git@64cc4c328f90da6d19316fcc1f6592acb903378c#egg=finam_x_hse_trade_ai_hack_trader
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mypy_extensions==1.1.0
narwhals==2.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.4.0
plotly==6.3.1
kaleido==0.2.1
prometheus-client==0.20.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
pluggy==1.6.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
requests==2.32.5
ruff==0.13.3
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
APScheduler==3.10.4
PyYAML==6.0.2
